Reorganized tree and wrapped the application with QueryClientProvider
