+++ package.json
@@ -18,6 +18,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-tooltip": "^1.0.7",
+    "@tanstack/react-query": "^5.40.0",
     "@types/lodash.debounce": "^4.0.9",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
@@ -31,6 +32,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.44.1",
+    "@tanstack/eslint-plugin-query": "^5.35.6",
     "@types/eslint": "^8.56.10",
     "@types/node": "^20.12.12",
     "@types/react": "^18.2.66",
