+++ src/api/const.ts
@@ -1,0 +1,1 @@
+export const API_ENDPOINT = `https://www.omdbapi.com/?apikey=${import.meta.env.VITE_APP_MOVIE_API_KEY}`;
