+++ src/api/fetchMovies.ts
@@ -1,4 +1,5 @@
-const API_ENDPOINT = `https://www.omdbapi.com/?apikey=${import.meta.env.VITE_APP_MOVIE_API_KEY}`;
+
+import { API_ENDPOINT } from "./const";
 
 export const fetchMovies = async (
   searchQuery: string,
@@ -21,23 +22,5 @@ export const fetchMovies = async (
     return { data: movies, totalResults: data.totalResults };
   } catch (error) {
     throw new Error((error as Error).message || "Failed to fetch movies");
-  }
-};
-
-export const fetchMovieDetails = async (id: string) => {
-
-  try {
-    const response = await fetch(`${API_ENDPOINT}&i=${id}`);
-    const data = await response.json();
-
-    if (!response.ok || data.Response !== "True") {
-      throw new Error(data.Error || "Failed to fetch movie details");
-    }
-
-    return data;
-  } catch (error) {
-    throw new Error(
-      (error as Error).message || "Failed to fetch movie details"
-    );
   }
 };
