+++ src/queryClient.ts
@@ -1,0 +1,5 @@
+import { QueryClient } from '@tanstack/react-query';
+
+const queryClient = new QueryClient();
+
+export default queryClient;
